Update Nextra to v4
