+++ .env.local
@@ -1,2 +1,1 @@
-PELICAN_PATH=/Users/clock/GolandProjects/pelican
-NODE_PATH=/Users/clock/WebstormProjects/docs/node_modules
+PELICAN_PATH=./pelican
