--- pages/placeholder.md
@@ -1,1 +1,0 @@
-This file is solely here so that the nextjs prechecks don't stop the compilation before the appropriate symlinks can be built in.
