+++ scripts/postdev.mjs
@@ -1,0 +1,4 @@
+import fs from 'fs';
+
+fs.rmSync('./app', { recursive: true, force: true });
+fs.rmSync('./public/pelican', { recursive: true, force: true });
